refactor: ai 관련 파일 모두 infra로 이동
