+++ src/main/java/com/greeenai/greeenai/domain/diary/service/DiaryService.java
@@ -2,10 +2,10 @@ package com.greeenai.greeenai.domain.diary.service;
 
 import static com.greeenai.greeenai.global.error.exception.ErrorCode.*;
 
-import com.greeenai.greeenai.domain.ai.dto.request.DiaryEntry;
-import com.greeenai.greeenai.domain.ai.dto.response.GenerateQuestionsResponse.GeneratedQuestion;
-import com.greeenai.greeenai.domain.ai.dto.response.GeneratedImage;
-import com.greeenai.greeenai.domain.ai.service.AIClient;
+import com.greeenai.greeenai.infra.ai.dto.request.DiaryEntry;
+import com.greeenai.greeenai.infra.ai.dto.response.GenerateQuestionsResponse.GeneratedQuestion;
+import com.greeenai.greeenai.infra.ai.dto.response.GeneratedImage;
+import com.greeenai.greeenai.infra.ai.client.AIClient;
 import com.greeenai.greeenai.domain.diary.domain.Diary;
 import com.greeenai.greeenai.domain.diary.domain.DiaryStatus;
 import com.greeenai.greeenai.domain.diary.domain.Option;
