+++ src/main/java/com/greeenai/greeenai/infra/ai/dto/request/DiaryEntry.java
@@ -1,4 +1,4 @@
-package com.greeenai.greeenai.domain.ai.dto.request;
+package com.greeenai.greeenai.infra.ai.dto.request;
 
 import com.greeenai.greeenai.domain.diary.domain.Question;
