+++ src/main/java/com/greeenai/greeenai/infra/ai/dto/response/GeneratedImage.java
@@ -1,4 +1,4 @@
-package com.greeenai.greeenai.domain.ai.dto.response;
+package com.greeenai.greeenai.infra.ai.dto.response;
 
 public record GeneratedImage(byte[] bytes, String contentType) {
     public static GeneratedImage of(byte[] bytes, String contentType) {
